Added many functions to interact with the camera
Added method to connect by specifying the camera Serial Number.
Added methods (set and get) for framerate, clock (frequency), and AOI size and position.
Also added method to find out the range and limits for the above parameters.
Also added method to change the AOI position in a fast way, and a method to verify such a method is supported.